+++ apps/meteor/tests/e2e/e2e-encryption.spec.ts
@@ -884,6 +884,8 @@ test.describe.serial('e2ee room setup', () => {
 
 		await page.goto('/home');
 
+		await page.waitForSelector('.main-content');
+
 		await expect(poHomeChannel.bannerSaveEncryptionPassword).toBeVisible();
 
 		const channelName = faker.string.uuid();