+++ packages/apps-engine/src/server/AppManager.ts
@@ -619,6 +619,8 @@ export class AppManager {
             return aff;
         }
 
+        app.getStorageItem()._id = created._id;
+
         this.apps.set(app.getID(), app);
         aff.setApp(app);
 
@@ -721,8 +723,13 @@ export class AppManager {
             implemented: result.implemented.getValues(),
             ...(old.marketplaceInfo && { marketplaceInfo: old.marketplaceInfo }),
             ...(old.sourcePath && { sourcePath: old.sourcePath }),
-            ...(permissionsGranted && { permissionsGranted }),
         };
+
+        if (!permissionsGranted) {
+            delete descriptor.permissionsGranted;
+        } else {
+            descriptor.permissionsGranted = permissionsGranted;
+        }
 
         try {
             descriptor.sourcePath = await this.appSourceStorage.update(descriptor, appPackage);